+++ src/port/electrum-chain-source.ts
@@ -100,7 +100,7 @@ export class WsElectrumChainSource implements ChainSource {
     }
   }
 
-  async waitForAddressReceivesTx(addr: string): Promise<void> {
+  waitForAddressReceivesTx(addr: string): Promise<void> {
     return new Promise((resolve, reject) => {
       this.subscribeScriptStatus(address.toOutputScript(addr), (_, status) => {
         if (status !== null) {
